Remove justified word breaks
The original text was justified, so some words are unnecessarily broken
with a dash e.g. "transla-tion". This PR removes the dashes and
subsequent line breaks.

I tried to be careful not to remove meaningful breaks (e.g. within a
word like "well-regulated"), but doubtlessly I made a few mistakes.