Fix index line ordering
The index in the source PDF is two columns. When Poppler parsed this, it
reads each line left to right instead of reading one column top to
bottom and then the other column.

That is, if the page read something like this:

A C
B D

Then the page would be read into HTML as:

A
C
B
D

This commit reorders the content into the proper order (A B C D).

The I did this is actually kind of interesting. I used Vim and began by
setting a "mark" (see https://vim.fandom.com/wiki/Using_marks) on the
first line (a) and second line (b) of the first column. In the above
example, mark 'a' is on line A and mark 'b' is on line B.

Mark 'a' represents the bottom of what we've already sorted from the
first column, and mark 'b' represents the next line that is mixed in
with column 2 that we need to move up to mark 'a'.

Then I ran a macro that worked like this (starting on mark 'a'):

	jump to mark 'b'
	cut the line
	move mark 'b' down two lines (to the next line from the second column)
	jump to mark 'a'
	paste the line (cut from 'b')
	move mark 'a' down one line (to the line we just pasted)

By repeating this, it pulls out each line from column one and pastes it
in order at the top. That leaves column 2 already sorted once the
content from column 1 is removed.

In the example above, we end up with:

A
B
C
D

Just as we'd like.